chore(appeals): move calculate due date to utils
