+++ appeals/api/src/server/endpoints/appeals/__tests__/appeals.test.js
@@ -15,10 +15,10 @@ import {
 	ERROR_MUST_BE_VALID_APPEAL_STATE,
 	ERROR_PAGENUMBER_AND_PAGESIZE_ARE_REQUIRED
 } from '@pins/appeals/constants/support.js';
-import { APPEAL_CASE_PROCEDURE, APPEAL_CASE_STATUS } from '@planning-inspectorate/data-model';
+import { APPEAL_CASE_STATUS } from '@planning-inspectorate/data-model';
 import { omit } from 'lodash-es';
 import { request } from '../../../app-test.js';
-import { getIdsOfReferencedAppeals, mapAppealToDueDate } from '../appeals.formatter.js';
+import { getIdsOfReferencedAppeals } from '../appeals.formatter.js';
 import { mapAppealStatuses } from '../appeals.service.js';
 const { databaseConnector } = await import('#utils/database-connector.js');
 
@@ -1419,255 +1419,6 @@ test('gets unassigned teams when assignedTeamId param is -1', async () => {
 		})
 	);
 	expect(response.status).toEqual(200);
-});
-describe('mapAppealToDueDate Tests', () => {
-	let mockAppeal = {
-		appealType: null,
-		id: 1,
-		address: null,
-		appellant: null,
-		agent: null,
-		lpa: null,
-		reference: 'APP/Q9999/D/21/33813',
-		caseExtensionDate: null,
-		caseCreatedDate: new Date('2023-01-01T00:00:00.000Z'),
-		appealStatus: [
-			{
-				id: 2648,
-				status: APPEAL_CASE_STATUS.READY_TO_START,
-				createdAt: new Date('2024-01-09T16:14:31.387Z'),
-				valid: true,
-				appealId: 496,
-				subStateMachineName: null,
-				compoundStateName: null
-			}
-		],
-		appealTimetable: null
-	};
-
-	beforeEach(() => {
-		mockAppeal = {
-			appellant: null,
-			agent: null,
-			address: null,
-			appealType: null,
-			id: 1,
-			lpa: null,
-			reference: 'APP/Q9999/D/21/33813',
-			caseExtensionDate: null,
-			caseCreatedDate: new Date('2023-01-01T00:00:00.000Z'),
-			appealStatus: [
-				{
-					id: 2648,
-					status: APPEAL_CASE_STATUS.READY_TO_START,
-					createdAt: new Date('2024-01-09T16:14:31.387Z'),
-					valid: true,
-					appealId: 496,
-					subStateMachineName: null,
-					compoundStateName: null
-				}
-			],
-			appealTimetable: null
-		};
-	});
-
-	test('maps STATE_TARGET_READY_TO_START status', async () => {
-		mockAppeal.appealStatus[0].status = APPEAL_CASE_STATUS.READY_TO_START;
-		// @ts-ignore
-		const dueDate = await mapAppealToDueDate(mockAppeal, 'Incomplete', new Date('2023-02-01'));
-		expect(dueDate).toEqual(new Date('2023-02-01'));
-	});
-
-	test('maps STATE_TARGET_READY_TO_START status with Incomplete status', async () => {
-		mockAppeal.appealStatus[0].status = APPEAL_CASE_STATUS.READY_TO_START;
-		const createdAtPlusFiveDate = new Date('2023-01-06T00:00:00.000Z');
-		// @ts-ignore
-		const dueDate = await mapAppealToDueDate(mockAppeal, '', null);
-		expect(dueDate).toEqual(createdAtPlusFiveDate);
-	});
-
-	test('maps STATE_TARGET_LPA_QUESTIONNAIRE_DUE', async () => {
-		mockAppeal.appealStatus[0].status = APPEAL_CASE_STATUS.LPA_QUESTIONNAIRE;
-
-		const createdAtPlusTenDate = new Date('2023-01-11T00:00:00.000Z');
-		// @ts-ignore
-		const dueDate = await mapAppealToDueDate(mockAppeal, '', null);
-		expect(dueDate).toEqual(createdAtPlusTenDate);
-	});
-
-	test('maps STATE_TARGET_LPA_QUESTIONNAIRE_DUE status with appealTimetable lpaQuestionnaireDueDate', async () => {
-		let mockAppealWithTimetable = {
-			...mockAppeal,
-			appealTimetable: {
-				id: 1262,
-				appealId: 523,
-				lpaQuestionnaireDueDate: new Date('2023-03-01T00:00:00.000Z'),
-				issueDeterminationDate: null,
-				lpaStatementDueDate: null
-			}
-		};
-		mockAppealWithTimetable.appealStatus[0].status = APPEAL_CASE_STATUS.LPA_QUESTIONNAIRE;
-
-		// @ts-ignore
-		const dueDate = await mapAppealToDueDate(mockAppealWithTimetable, '', null);
-		expect(dueDate).toEqual(new Date('2023-03-01T00:00:00.000Z'));
-	});
-
-	test('maps STATE_TARGET_ASSIGN_CASE_OFFICER', async () => {
-		mockAppeal.appealStatus[0].status = APPEAL_CASE_STATUS.ASSIGN_CASE_OFFICER;
-
-		const createdAtPlusFifteenDate = new Date('2023-01-16T00:00:00.000Z');
-		// @ts-ignore
-		const dueDate = await mapAppealToDueDate(mockAppeal, '', null);
-		expect(dueDate).toEqual(createdAtPlusFifteenDate);
-	});
-
-	test('maps STATE_TARGET_ISSUE_DETERMINATION when site visit available', async () => {
-		mockAppeal.appealStatus[0].status = APPEAL_CASE_STATUS.ISSUE_DETERMINATION;
-		mockAppeal.siteVisit = { visitDate: new Date('2023-02-01T00:00:00.000Z') };
-		const createdAtPlusFortyBusinessDays = new Date('2023-03-29T00:00:00.000Z');
-		// @ts-ignore
-		const dueDate = await mapAppealToDueDate(mockAppeal, '', null);
-		expect(dueDate).toEqual(createdAtPlusFortyBusinessDays);
-	});
-
-	test('maps STATE_TARGET_ISSUE_DETERMINATION when site visit not available', async () => {
-		mockAppeal.appealStatus[0].status = APPEAL_CASE_STATUS.ISSUE_DETERMINATION;
-
-		const createdAtPlusThirtyBusinessDays = new Date('2025-07-04T00:00:00.000Z');
-		mockAppeal.caseCreatedDate = new Date('2025-05-22T00:00:00.000Z');
-
-		// @ts-ignore
-		const dueDate = await mapAppealToDueDate(mockAppeal, '', null);
-		expect(dueDate).toEqual(createdAtPlusThirtyBusinessDays);
-	});
-
-	test('maps STATE_TARGET_ISSUE_DETERMINATION status with appealTimetable issueDeterminationDate', async () => {
-		let mockAppealWithTimetable = {
-			...mockAppeal,
-			appealTimetable: {
-				id: 1262,
-				appealId: 523,
-				lpaQuestionnaireDueDate: null,
-				issueDeterminationDate: new Date('2023-03-01T00:00:00.000Z'),
-				lpaStatementDueDate: null
-			}
-		};
-		mockAppealWithTimetable.appealStatus[0].status = APPEAL_CASE_STATUS.ISSUE_DETERMINATION;
-		mockAppealWithTimetable.caseCreatedDate = new Date('2025-05-22T00:00:00.000Z');
-		// @ts-ignore
-		const dueDate = await mapAppealToDueDate(mockAppealWithTimetable, '', null);
-		// dueDate.setDate(dueDate.getDate());
-		expect(dueDate).toEqual(new Date('2025-07-04T00:00:00.000Z'));
-	});
-
-	test('maps STATE_TARGET_STATEMENT_REVIEW', async () => {
-		mockAppeal.appealStatus[0].status = APPEAL_CASE_STATUS.STATEMENTS;
-
-		const createdAtPlusFiftyFiveDate = new Date('2023-02-25T00:00:00.000Z');
-		// @ts-ignore
-		const dueDate = await mapAppealToDueDate(mockAppeal, '', null);
-		expect(dueDate).toEqual(createdAtPlusFiftyFiveDate);
-	});
-
-	test('maps STATE_TARGET_STATEMENT_REVIEW status with appealTimetable lpaStatementDueDate', async () => {
-		let mockAppealWithTimetable = {
-			...mockAppeal,
-			appealTimetable: {
-				id: 1262,
-				appealId: 523,
-				lpaQuestionnaireDueDate: null,
-				issueDeterminationDate: null,
-				lpaStatementDueDate: new Date('2023-03-01T00:00:00.000Z'),
-				resubmitAppealTypeDate: null
-			}
-		};
-		mockAppealWithTimetable.appealStatus[0].status = APPEAL_CASE_STATUS.STATEMENTS;
-
-		// @ts-ignore
-		const dueDate = await mapAppealToDueDate(mockAppealWithTimetable, '', null);
-		expect(dueDate).toEqual(new Date('2023-03-01T00:00:00.000Z'));
-	});
-
-	test('maps STATE_TARGET_FINAL_COMMENT_REVIEW', async () => {
-		mockAppeal.appealStatus[0].status = APPEAL_CASE_STATUS.FINAL_COMMENTS;
-
-		const createdAtPlusSixtyDate = new Date('2023-03-02T00:00:00.000Z');
-		// @ts-ignore
-		const dueDate = await mapAppealToDueDate(mockAppeal, '', null);
-		expect(dueDate).toEqual(createdAtPlusSixtyDate);
-	});
-
-	test('handles STATE_TARGET_AWAITING_SITE_VISIT', async () => {
-		mockAppeal.appealStatus[0].status = APPEAL_CASE_STATUS.AWAITING_EVENT;
-		mockAppeal.siteVisit = { visitDate: new Date('2023-02-01T00:00:00.000Z') };
-		mockAppeal.procedureType = {
-			key: APPEAL_CASE_PROCEDURE.WRITTEN
-		};
-
-		// @ts-ignore
-		const dueDate = await mapAppealToDueDate(mockAppeal, '', null);
-		expect(dueDate).toEqual(mockAppeal.siteVisit.visitDate);
-	});
-
-	test('handles STATE_TARGET_AWAITING_HEARING', async () => {
-		mockAppeal.appealStatus[0].status = APPEAL_CASE_STATUS.AWAITING_EVENT;
-		mockAppeal.hearing = { hearingStartTime: new Date('2023-02-01T00:00:00.000Z') };
-		mockAppeal.procedureType = {
-			key: APPEAL_CASE_PROCEDURE.HEARING
-		};
-
-		// @ts-ignore
-		const dueDate = await mapAppealToDueDate(mockAppeal, '', null);
-		expect(dueDate).toEqual(mockAppeal.hearing.hearingStartTime);
-	});
-
-	describe('handles STATE_TARGET_SITE_VISIT', () => {
-		let mockAppealWithTimetable = {};
-
-		beforeEach(() => {
-			mockAppealWithTimetable = {
-				...mockAppeal,
-				appealTimetable: {
-					id: 1262,
-					appealId: 523,
-					lpaQuestionnaireDueDate: new Date('2023-03-01T00:00:00.000Z')
-				}
-			};
-			mockAppealWithTimetable.appealStatus[0].status = APPEAL_CASE_STATUS.EVENT;
-		});
-
-		test('when final comments due date does not exist', async () => {
-			// @ts-ignore
-			const dueDate = await mapAppealToDueDate(mockAppealWithTimetable, '', null);
-			expect(dueDate).toEqual(mockAppealWithTimetable.appealTimetable.lpaQuestionnaireDueDate);
-		});
-
-		test('when final comments due date does exist', async () => {
-			// @ts-ignore
-			mockAppealWithTimetable.appealTimetable.finalCommentsDueDate = new Date(
-				'2023-03-22T00:00:00.000Z'
-			);
-			const dueDate = await mapAppealToDueDate(mockAppealWithTimetable, '', null);
-			expect(dueDate).toEqual(mockAppealWithTimetable.appealTimetable.finalCommentsDueDate);
-		});
-	});
-
-	test('handles STATE_TARGET_COMPLETE', async () => {
-		mockAppeal.appealStatus[0].status = APPEAL_CASE_STATUS.COMPLETE;
-
-		// @ts-ignore
-		const dueDate = await mapAppealToDueDate(mockAppeal, '', null, {});
-		expect(dueDate).toBeNull();
-	});
-
-	test('handles unexpected status (default case)', async () => {
-		mockAppeal.appealStatus[0].status = 'unexpected_status';
-
-		// @ts-ignore
-		const dueDate = await mapAppealToDueDate(mockAppeal, '', null);
-		expect(dueDate).toBeUndefined();
-	});
 });
 
 describe('mapAppealStatuses Tests', () => {
